requirements.txt
Flask==2.3.2
mysql-connector-python==8.3.0
matplotlib==3.8.0
requests==2.31.0
Jinja2==3.1.2
numpy==1.25.0